Post file upload functionality

+++ routes/web.php
@@ -31,40 +31,39 @@ Route::view('/', 'welcome');
 * Admin routes
 */
 
-Route::prefix('admin')->group(function() {
+Route::prefix('admin')->group(function () {
     /*
     * Auth
     */
     Route::view('/', 'admin.pages.dashboard')->middleware('auth')->name('admin.dashboard');
 
-	Route::get('/login', [LoginController::class, 'showLoginForm'])->name('login');
-	Route::post('/login', [LoginController::class, 'login']);
-	Route::get('/wyloguj', [LoginController::class, 'logout'])->name('admin.logout');
+    Route::get('/login', [LoginController::class, 'showLoginForm'])->name('login');
+    Route::post('/login', [LoginController::class, 'login']);
+    Route::get('/wyloguj', [LoginController::class, 'logout'])->name('admin.logout');
 
-	Route::get('/przypomnienie', [ForgotPasswordController::class, 'showLinkRequestForm'])->name('password.request');
-	Route::post('/email', [ForgotPasswordController::class, 'sendResetLinkEmail'])->name('password.email');
+    Route::get('/przypomnienie', [ForgotPasswordController::class, 'showLinkRequestForm'])->name('password.request');
+    Route::post('/email', [ForgotPasswordController::class, 'sendResetLinkEmail'])->name('password.email');
 
-	Route::get('/resetowanie/{token}', [ResetPasswordController::class, 'showResetForm'])->name('password.reset');
-	Route::post('/resetowanie/', [ResetPasswordController::class, 'reset'])->name('password.update');
+    Route::get('/resetowanie/{token}', [ResetPasswordController::class, 'showResetForm'])->name('password.reset');
+    Route::post('/resetowanie/', [ResetPasswordController::class, 'reset'])->name('password.update');
 
-	Route::get('/kategorie', [CategoryController::class, 'index'])->middleware('auth')->name('admin.categories');
-	Route::get('/kategorie/formularz', [CategoryController::class, 'form'])->middleware('auth')->name('admin.categories.form');
-	Route::get('/kategorie/formularz/{categoryId}', [CategoryController::class, 'form'])->middleware('auth')->name('admin.categories.form.edit');
-	Route::post('/kategorie/formularz', [CategoryController::class, 'store'])->middleware('auth')->name('admin.categories.form.save');
-	Route::post('/kategorie/formularz/{categoryId}', [CategoryController::class, 'store'])->middleware('auth')->name('admin.categories.form.editSave');
+    Route::get('/kategorie', [CategoryController::class, 'index'])->middleware('auth')->name('admin.categories');
+    Route::get('/kategorie/formularz', [CategoryController::class, 'form'])->middleware('auth')->name('admin.categories.form');
+    Route::get('/kategorie/formularz/{categoryId}', [CategoryController::class, 'form'])->middleware('auth')->name('admin.categories.form.edit');
+    Route::post('/kategorie/formularz', [CategoryController::class, 'store'])->middleware('auth')->name('admin.categories.form.save');
+    Route::post('/kategorie/formularz/{categoryId}', [CategoryController::class, 'store'])->middleware('auth')->name('admin.categories.form.editSave');
 
-	Route::get('/ustawienia', [SettingsController::class, 'index'])->middleware('auth')->name('admin.settings');
     /*
     * News
     */
     Route::get('/aktualnosci', [PostController::class, 'index'])->middleware('auth')->name('admin.posts');
 
-    Route::get('/aktualnosci/utworz', [PostController::class, 'create'])->name('admin.posts.create');
-    Route::post('/aktualnosci/utworz', [PostController::class, 'store']);
+    Route::get('/aktualnosci/utworz', [PostController::class, 'create'])->middleware('auth')->name('admin.posts.create');
+    Route::post('/aktualnosci/utworz', [PostController::class, 'store'])->middleware('auth');
 
-    Route::get('/aktualnosci/{id}', [PostController::class, 'edit'])->name('admin.posts.edit');
-    Route::put('/aktualnosci/{id}', [PostController::class, 'update']);
-    Route::delete('/aktualnosci/{id}', [PostController::class, 'destroy'])->name('admin.posts.delete');
+    Route::get('/aktualnosci/{id}', [PostController::class, 'edit'])->middleware('auth')->name('admin.posts.edit');
+    Route::put('/aktualnosci/{id}', [PostController::class, 'update'])->middleware('auth');
+    Route::delete('/aktualnosci/{id}', [PostController::class, 'destroy'])->middleware('auth')->name('admin.posts.delete');
 
     /*
     * Settings
